Separate the form from the AddMeasurePage Component
